+++ core/src/main/java/org/ode4j/ode/internal/ErrorHdl.java
@@ -25,8 +25,11 @@
 package org.ode4j.ode.internal;
 
 import org.ode4j.ode.internal.ErrorHandler.dMessageFunction;
+import org.ode4j.ode.internal.cpp4j.Cstdio;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.IllegalFormatException;
 
 
 /**
@@ -115,14 +118,12 @@ public class ErrorHdl {
 	 * @param ap objects 
 	 */
 	public static void dError (int num, String msg, Object ... ap) {
-		//  va_list ap;
-		//  va_start (ap,msg);
 		if (error_function != null) {
-			error_function.call (num,msg,ap);
+			error_function.call(num, msg, ap);
 		} else {
-			logger.error("ODE Error " + num + ": " + msg, ap);
+			logger.error("ODE Error {}: {}", num, format(msg, ap));
 		}
-		throw new RuntimeException("#"+num + ": " + msg);
+		throw new RuntimeException("#" + num + ": " + format(msg, ap));
 		//exit (1);
 	}
 
@@ -134,16 +135,15 @@ public class ErrorHdl {
 	 * @param ap objects 
 	 */
 	public static void dDebug (int num, String msg, Object ... ap) {
-		//  va_list ap;
-		//  va_start (ap,msg);
 		if (debug_function != null) {
-			debug_function.call (num,msg,ap);
+			debug_function.call(num, msg, ap);
 		} else {
-			logger.debug("ODE INTERNAL ERROR " + " " + num + ": " + msg, ap);
+			if (logger.isDebugEnabled()) {
+				logger.debug("ODE INTERNAL ERROR {}: {}", num, format(msg, ap));
+			}
 		}
-		// *((char *)0) = 0;   ... commit SEGVicide
 		//abort();
-		throw new RuntimeException("#"+num + ": " + String.format(msg, ap));
+		throw new RuntimeException("#" + num + ": " + format(msg, ap));
 	}
 
 
@@ -154,13 +154,28 @@ public class ErrorHdl {
 	 * @param ap objects 
 	 */
 	public static void dMessage (int num, String msg, Object ... ap) {
-		//  va_list ap;
-		//  va_start (ap,msg);
-		if (message_function!=null) {
-			message_function.call (num,msg,ap);
+		if (message_function != null) {
+			message_function.call(num, msg, ap);
 		} else {
 			//printMessage (num,"ODE Message",msg,ap);
-			logger.info("ODE Message " + num + ": " + msg, ap);
+			if (logger.isInfoEnabled()) {
+				logger.info("ODE Message {}: {}", num, format(msg, ap));
+			}
+		}
+	}
+
+	private static String format(String msg, Object... ap) {
+		try {
+			return String.format(msg, ap);
+		} catch (IllegalFormatException e) {
+			// Ensure that we have an error message even if the formatting is bad
+			StringBuilder sb = new StringBuilder();
+			sb.append(msg).append(" ");
+			for (Object o: ap) {
+				sb.append(o).append(",");
+			}
+			sb.append(e.getMessage());
+			return sb.toString();
 		}
 	}